keep working on AI, rework MoveComPlatform()
I have cancelled a feature that AI tends to catch the ball by middle of
the platform.
Collision point was in between 2/5 and 3/5 of platform width.
Now MoveComPlatform() is also used to move platform to wrong position,
which may have the case like move to position 0, it's definatly not in
between for platform 2/5 and 3/5.